refactor: trace
